backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
click==8.2.0
colorama==0.4.6
fastapi==0.115.12
flask-cors==5.0.1
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
Werkzeug==3.1.3
pymongo==4.13.0
Pillow==11.2.1
